Обновленная версия
