+++ src/Java2/Lesson1/Let.java
@@ -1,0 +1,5 @@
+package Java2.Lesson1;
+
+interface Let {
+    void info();
+}
